rewrite a few things
