--- src/core/screens/index.ts
@@ -1,3 +1,0 @@
-export * as Plugins from './Plugins';
-export * as Themes from './Themes';
-export * as Enmity from './Enmity';
