--- src/modules/metro/react.ts
@@ -1,3 +1,0 @@
-import { common } from '@metro';
-
-export default common.React as typeof import('react-types');
